Add a script to promote docker images
Add a script to promote docker images between environments and
also to push the images to GCR. It's easier to tag all images with
the GCR container registry at this point in time, so this commit
also changes the tag format to include that information. Moreover,
the image name and versions were starting to get a little unwieldy
so this commit cleans those up and gather them into some settings
as well.