Make waiting room use queryParams for roomId and playerId
Using the HIT Id and the Assignment Id query params to create the
room and players means that we can prevent workers from playing
themselves (i.e., taking HITs for all the players) by having the
number of assignments be the number of players. Additionally, the
games can be initialized to the waiting room without worrying about
coordinating the URLs to be distinct.