Refactor code to remove boilerplate controller callbacks
Many of the controller callbacks had the form of wrapping some
method from the model, updating the view, and then updating the
server. All of these methods can be abstracted by the newly
written 'dispatchAction' method. This commit refactors the code
to replace the old callbacks with the dispatchAction method.