Move askerId and answererId up to the Game class
askerId and answererId can change throughout the coarse of a game,
and thus represent part of the game's state, so they shouldn't be
on the rounds. In the future, if a player abandons then they could
be replaced by a new asker or answerer, so to have the askerId and
the answererId on a per round basis seems misleading.