Use server defined routing for pairing players with games
This commit changes the models and views on both the client and
the server so that rather than hard coding player-game matches in
URLs, the server can dynamically pair players with games. This prevents
some of the issues with Turkers picking up a HIT and not working it. It
will also enable the server to replace inactive or abandoned players in
the future.