Major refactor to support making players go inactive
We need to be able to put players in an inactive state and then
reactivate them when they return to their browser. In order to
implement this well, we have to separate out the player and
game room models with a sensible way or handling the complicated
routing logic.

Additionally, we want to merge the waiting room and the game room
since handling them together means that if a player comes back to
a game they're currently in or hits refresh on their browser then
they don't have to go through the waiting room. Similarly, if we
make a player inactive, we don't need to route them to a new page,
and other similar issues.