Make answers to questions 5 valued
Initial runs of the twenty questions procedure show that having
only two values for the answers to questions leads to ambiguity in
enough cases that the human baseline for the generated data is only
around 85% accuracy. An example question that's ambiguous would be:

    is a car blue?

Cars can be blue, but it's not as if most cars are blue. Moving the
game to have 5 valued answers requires a separate model for questions
and guesses, since guesses are still either true or false. This commit
changes the backend and the frontend to reflect the five valued answers
for questions and to split out the model of guesses from the model of
questions and answers.