Make finish game a no-op
It complicates the logic to have a player deleted in two places
(when they finish the game and when they disconnect). Also, if a
player doesn't successfully post to MTurk to finish their HIT, we
want them to be able to post a second time. This change also fixes
a bug in disconnect where a player would get deleted when finishing
a game and then disconnect would throw a key error when looking for
the player's match.