Create multiple domains and servers for twenty questions
We need to scale horizontally to create the questions faster. In
particular, we'll do this by creating multiple servers and HITs
pointing the turkers to specific servers to balance the load rather
than using sticky sessions which could lead to turkers losing the
session and the connection with their game.