Update the subjects list and fix duplicates from the first list
The prior list had some duplicates in it due to an issue in how
it was sampled (with replacement) and duplicates in the list of
common nouns. A new sample of the remaining nouns was taken after
more aggressive deduplication of the list of common nouns in
combination with sampling without replacement.