Add a READINGINSTRUCTIONS player status and refactor instructions
We want to improve the instructions to encourage more diversity
from the turkers; however, this makes the instructions more complex
and take more time to read. So, we need to add a specific state that
provides turkers time to read the instructions. Also, we can reduce
repetition by factoring out the instructions into their own component.